products finished, fixed clients/settings styles

+++ db/migrate/20160713220956_create_products.rb
@@ -5,10 +5,12 @@ class CreateProducts < ActiveRecord::Migration
       t.text :description
       t.decimal :unitprice, precision: 11, scale: 2 # Big numbers with just two decimal places
       t.decimal :total, precision: 11, scale: 2 # Big numbers with just two decimal places
-      t.float :btw, precision: 3, scale: 2 # 1.0, 0.5, etc.
-      t.float :discount, precision: 3, scale: 2  # 1.0, 0.5, etc.
+      t.decimal :btw, precision: 3, scale: 2 # 1.0, 0.5, etc.
+      t.decimal :discount, precision: 3, scale: 2  # 1.0, 0.5, etc.
 
       t.timestamps null: false
     end
   end
 end
+
+
